+++ packages/framework/src/hooks/useTemplateData.ts
@@ -4,9 +4,10 @@ import { isString, map } from "lodash-es";
 import isEqual from "react-fast-compare";
 import { selectAtom } from "jotai/utils";
 import type { ScreenContextDefinition } from "../state";
-import { screenAtom } from "../state";
+import { screenAtom } from "../state/screen";
 import { evaluate } from "../evaluate";
 import type { Expression } from "../shared/common";
+import { createStorageApi } from "./useEnsembleStorage";
 
 export type TemplateData = object | unknown[];
 export interface TemplateDataProps {
@@ -39,7 +40,11 @@ export const useTemplateData = ({
             return data;
           }
           try {
-            return evaluate(screenContext, String(data));
+            return evaluate(screenContext, String(data), {
+              ensemble: {
+                storage: createStorageApi(screenContext.storage),
+              },
+            });
           } catch (e) {
             return {};
           }