+++ packages/framework/src/index.ts
@@ -4,4 +4,5 @@ export * from "./parser";
 export * from "./hooks";
 export * from "./data";
 export * from "./state";
+export * from "./storage";
 export { evaluate } from "./evaluate";