+++ packages/runtime/src/runtime/customWidget.tsx
@@ -1,4 +1,7 @@
-import { CustomScopeProvider } from "@ensembleui/react-framework";
+import {
+  CustomScopeProvider,
+  useRegisterBindings,
+} from "@ensembleui/react-framework";
 import type { CustomWidgetModel } from "@ensembleui/react-framework";
 import React from "react";
 import { EnsembleRuntime } from "./runtime";
@@ -11,8 +14,11 @@ export const createCustomWidget = (
   model: CustomWidgetModel,
 ): React.FC<CustomWidgetProps> => {
   const CustomWidget: React.FC<CustomWidgetProps> = ({ inputs }) => {
+    const { values } = useRegisterBindings<Record<string, unknown>>({
+      ...inputs,
+    });
     return (
-      <CustomScopeProvider value={inputs}>
+      <CustomScopeProvider value={values ?? inputs}>
         {EnsembleRuntime.render([model.body])}
       </CustomScopeProvider>
     );