+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @ensembleui/react-runtime
 
+## 0.0.13
+
+### Patch Changes
+
+- 84deb25: Normalize dependencies
+
+## 0.0.12
+
+### Patch Changes
+
+- bf12a2e: Fix runtime package dependency
+
+## 0.0.11
+
+### Patch Changes
+
+- 67bffb7: Skip screen height calculation if modal
+- 284be43: Only store state for widgets with explicit ids
+- b70f000: Skip navigating to selected screen if path is not empty
+- 51bfb7c: Register custom widget inputs as bindings
+- e238827: Improve chart config evaluation
+- dbd93e5: Trigger binding updates on storage change
+
 ## 0.0.10
 
 ### Patch Changes